separeted code on modules, made a webpack build
