+++ public/index.html
@@ -81,6 +81,7 @@
 <template id="card-description-template">
   <p class="film-card__description">A priest with a haunted past and a novice on the threshold of her final vows are sent by the Vatican to investigate the death of a young nun in Romania and confront a malevolent force in the form of a demonic nun.</p>
 </template>
-<script src="../src/main.js"></script>
+<!--<script src="../src/main.js"></script>-->
+<script src="./bundle.js"></script>
 </body>
 </html>
