+++ src/main.js
@@ -20,6 +20,9 @@
 8. Добавьте обработчик для переключения фильтров. При их переключении удаляйте все ранее созданные фильмы
  и добавляйте случайное количество новых.*/
 
+import createFilter from './create-filter.js';
+import createCard from './create-card.js';
+
 const filtersContainer = document.querySelector(`.main-navigation`);
 const CardsContainer = document.querySelector(`.films`);
 const mainCardsContainer = CardsContainer.querySelector(`.films-list__container`);
@@ -44,16 +47,6 @@ const CardsNumber = {
   EXTRA: 2
 };
 
-const CARD_PROPERTIES = [
-  {name: `title`, value: `Incredibles 2`},
-  {name: `rating`, value: `9.8`},
-  {name: `year`, value: 2018},
-  {name: `duration`, value: `1h 13m`},
-  {name: `genre`, value: `Comedy`},
-  {name: `poster`, value: `./images/posters/accused.jpg`},
-  {name: `comments`, value: `13 comments`}
-];
-
 const FILTERS_NAMES = [`All movies`, `Watchlist`, `History`, `Favorites`];
 
 /**
@@ -63,62 +56,6 @@ const FILTERS_NAMES = [`All movies`, `Watchlist`, `History`, `Favorites`];
  * @return {number}
  */
 const getRandomNumber = (min, max) => Math.floor(Math.random() * (max - min + 1)) + min;
-
-/**
- * returns markup of filter element template as document fragment
- * @param {string} name
- * @param {number} number
- * @return {DocumentFragment}
- */
-const createFilter = (name, number) => {
-  const filter = document.querySelector(`#filter-template`).content.cloneNode(true);
-  const fragment = document.createDocumentFragment();
-  const parentDiv = filter.querySelector(`.main-navigation__item`);
-  const numberOfFilms = filter.querySelector(`.main-navigation__item-count`);
-  const href = name.toLowerCase().split(` `)[0];
-
-  if (href === `all`) {
-    parentDiv.textContent = ``;
-    parentDiv.classList.add(`main-navigation__item--active`);
-  } else {
-    numberOfFilms.textContent = number;
-  }
-
-  parentDiv.insertAdjacentText(`afterbegin`, name);
-  parentDiv.setAttribute(`href`, `#${href}`);
-
-  fragment.appendChild(filter);
-
-  return fragment;
-};
-
-/**
- * returns markup of card element template as document fragment
- * @param {boolean} hasControls
- * @return {DocumentFragment}
- */
-const createCard = (hasControls) => {
-  const card = document.querySelector(`#card-template`).content.cloneNode(true);
-  const fragment = document.createDocumentFragment();
-  const parentDiv = card.querySelector(`.film-card`);
-
-  CARD_PROPERTIES.forEach((property) => {
-    card.querySelector(`.film-card__${property.name}`).textContent = property.value;
-  });
-
-  if (hasControls) {
-    const controls = document.querySelector(`#card-controls-template`).content.cloneNode(true);
-    const description = document.querySelector(`#card-description-template`).content.cloneNode(true);
-    parentDiv.appendChild(controls);
-    parentDiv.insertBefore(description, card.querySelector(`img`));
-  } else {
-    parentDiv.classList.add(`film-card--no-controls`);
-  }
-
-  fragment.appendChild(card);
-
-  return fragment;
-};
 
 /**
  * inserts the resulting nodes (cards) into the DOM tree
